requirements.txt
fastapi
uvicorn
langchain
langchain-community
langchain-huggingface
faiss-cpu
sentence-transformers
pydantic
python-dotenv
pypdf
transformers
torch
accelerate
bitsandbytes
streamlit
PyMuPDF
openai
google-generativeai
requests
